Modification de la couleur de la porte de sortie des ghost et réarrangement des position des ghost

See merge request hem/pacman!5

+++ src/main/java/model/MazeState.java
@@ -45,7 +45,7 @@ public final class MazeState {
         return critters;
     }
 
-    public double getWidth() {
+    public int getWidth() {
         return width;
     }
